Split out ReplacedTexture into its own file.
